+++ handlers/movies_watchnow_handler.go
@@ -25,7 +25,7 @@ func (MoviesWatchNowHandler) Handle(options *str.Options, client *internal.Clien
 	printer.Println("Returns streaming and watch now sources for a movie in the requested country (limited access).")
 	opts := uri.ListOptions{Extended: options.ExtendedInfo, Links: options.Links}
 	result, resp, err := client.Movies.GetMovieWatchNow(client.BuildCtxFromOptions(options), &options.InternalID, &options.Country, &opts)
-	if err = watchNowError(consts.WatchNow, consts.Movie, options, resp, err); err != nil {
+	if err = watchNowError(consts.WatchNow, consts.Movie, options.InternalID, resp, err); err != nil {
 		return err
 	}