+++ handlers/shows_justwatch_links_handler.go
@@ -23,7 +23,7 @@ func (ShowsJustwatchLinksHandler) Handle(options *str.Options, client *internal.
 
 	printer.Println("Returns JustWatch links for a show in the requested country (limited access).")
 	result, resp, err := client.Shows.GetShowJustwatchLinks(client.BuildCtxFromOptions(options), &options.InternalID, &options.Country)
-	if err = watchNowError(consts.JustwatchLinks, consts.Show, options, resp, err); err != nil {
+	if err = watchNowError(consts.JustwatchLinks, consts.Show, options.InternalID, resp, err); err != nil {
 		return err
 	}